// src/model/assetResponseData.ts
/**
 * Shotstack
 * Official Typescript SDK for the Shotstack Cloud Video Editing API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AssetResponseAttributes } from './assetResponseAttributes';

/**
* The type of resource (an asset) and attributes of the asset.
*/
export class AssetResponseData {
    /**
    * The type of resource, in this case it is an assets.
    */
    'type': string;
    'attributes': AssetResponseAttributes;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "AssetResponseAttributes"
        }    ];

    static getAttributeTypeMap() {
        return AssetResponseData.attributeTypeMap;
    }
}