// src/model/flipTransformation.ts
/**
 * Shotstack
 * Official Typescript SDK for the Shotstack Cloud Video Editing API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Flip a clip vertically or horizontally. Acts as a mirror effect of the clip along the selected plane.
*/
export class FlipTransformation {
    /**
    * Flip a clip horizontally.
    */
    'horizontal'?: boolean;
    /**
    * Flip a clip vertically.
    */
    'vertical'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "horizontal",
            "baseName": "horizontal",
            "type": "boolean"
        },
        {
            "name": "vertical",
            "baseName": "vertical",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return FlipTransformation.attributeTypeMap;
    }
}