// src/model/generatedAssetErrorResponse.ts
/**
 * Shotstack
 * Official PHP SDK for the Shotstack Cloud Video Editing API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GeneratedAssetErrorResponseData } from './generatedAssetErrorResponseData';

/**
* Error response data for validation and other errors returned by the Create API.
*/
export class GeneratedAssetErrorResponse {
    /**
    * An array of errors.
    */
    'errors': Array<GeneratedAssetErrorResponseData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "errors",
            "baseName": "errors",
            "type": "Array<GeneratedAssetErrorResponseData>"
        }    ];

    static getAttributeTypeMap() {
        return GeneratedAssetErrorResponse.attributeTypeMap;
    }
}