// src/model/generatedAssetResponse.ts
/**
 * Shotstack
 * Official Typescript SDK for the Shotstack Cloud Video Editing API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GeneratedAssetResponseData } from './generatedAssetResponseData';

/**
* The response returned by the Create API [generate asset](#generate-asset) and [get generated asset](#get-generated-asset) requests. Includes status and details of the generated asset. The response follows the [json:api](https://jsonapi.org/) specification.
*/
export class GeneratedAssetResponse {
    'data': GeneratedAssetResponseData;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "GeneratedAssetResponseData"
        }    ];

    static getAttributeTypeMap() {
        return GeneratedAssetResponse.attributeTypeMap;
    }
}