// src/model/generatedAssetResponseData.ts
/**
 * Shotstack
 * Official PHP SDK for the Shotstack Cloud Video Editing API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GeneratedAssetResponseAttributes } from './generatedAssetResponseAttributes';

/**
* The type of resource (an asset), it\'s id and attributes of the generated file.
*/
export class GeneratedAssetResponseData {
    /**
    * The type of resource, in this case it is an asset.
    */
    'type': string;
    /**
    * The generated asset id.
    */
    'id': string;
    'attributes': GeneratedAssetResponseAttributes;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "GeneratedAssetResponseAttributes"
        }    ];

    static getAttributeTypeMap() {
        return GeneratedAssetResponseData.attributeTypeMap;
    }
}